+++ resource/ebpf_program/handler/lcp.py
@@ -1,9 +1,13 @@
+from resource.base.handler.lcp import LCP as Base_LCP
+
+from requests import delete as delete_req
+from requests import post as post_req
+from requests import put as put_req
+from requests.auth import HTTPBasicAuth as HTTP_Basic_Auth
+
 from document.ebpf_program.catalog import eBPF_Program_Catalog_Document
 from document.exec_env import Exec_Env_Document
 from lib.response import Unprocessable_Entity_Response
-from requests import post as post_req, put as put_req, delete as delete_req
-from requests.auth import HTTPBasicAuth as HTTP_Basic_Auth
-from resource.base.handler.lcp import LCP as Base_LCP
 from utils.log import Log
 from utils.sequence import wrap
 
@@ -62,7 +66,8 @@ class LCP(Base_LCP):
     def __apply(self, instance, exec_env, caller, data):
         h, p = exec_env.hostname, exec_env.lcp.port
         u, ps = exec_env.lcp.username, exec_env.lcp.password
-        resp_caller = caller(f'http://{h}:{p}/code', auth=HTTP_Basic_Auth(u, ps),
+        schema = 'https' if exec_env.lcp.https else 'http'
+        resp_caller = caller(f'{schema}://{h}:{p}/code', auth=HTTP_Basic_Auth(u, ps),
                              json=data(self.catalog))
         if resp_caller.content:
             try: